+++ src/features/rockets/rockets.ts
@@ -5,26 +5,17 @@ import { SOFTCAPS, createSoftcap } from "@/util/softcapped";
 import { computed } from "@vue/reactivity";
 import Decimal from "break_eternity.js";
 import { basics } from "../basics/basics";
+import { rocketFuel } from "../rocketFuel/rocketFuel";
 
 interface RocketData {
   startingReq: Decimal;
   gainMult: Decimal;
   resetGain: Decimal;
   nextAt: Decimal;
-  rfCost: Decimal;
-  rfEff1: Decimal;
-  rfEff2: Decimal;
-  rfEff3: Decimal;
-  rfEff4: Decimal;
   effExp: Decimal;
   effMult: Decimal;
   maxVelMult: Decimal;
   accMult: Decimal;
-}
-
-interface RocketActions {
-  rocketUp: () => void;
-  fuelUp: () => void;
 }
 
 export const ROCKET_EFF_EXP_SOFTCAP = createSoftcap({
@@ -32,20 +23,17 @@ export const ROCKET_EFF_EXP_SOFTCAP = createSoftcap({
   start: 3.5,
 });
 
-export const rockets: Feature<RocketData, RocketActions> = addFeature(
-  "rockets",
-  {
+export const rockets: Feature<RocketData, { rocketUp: () => void }> =
+  addFeature("rockets", {
     unl: {
       reached: computed(() => Decimal.gte(player.distance, 1e6)),
-      desc: computed(
-        () => "Reach " + formatDistance(1e6) + " to unlock Rockets."
-      ),
+      desc: computed(() => `Reach ${formatDistance(1e6)} to unlock Rockets.`),
     },
 
     data: {
-      startingReq: computed(() => Decimal.div(1e6, rockets.data.rfEff2.value)),
+      startingReq: computed(() => Decimal.div(1e6, rocketFuel.data.eff2.value)),
 
-      gainMult: computed(() => rockets.data.rfEff3.value),
+      gainMult: computed(() => rocketFuel.data.eff3.value),
 
       resetGain: computed(() =>
         Decimal.div(player.distance, rockets.data.startingReq.value)
@@ -61,25 +49,15 @@ export const rockets: Feature<RocketData, RocketActions> = addFeature(
           .times(rockets.data.startingReq.value)
       ),
 
-      rfCost: computed(() =>
-        Decimal.pow(1.6, Decimal.pow(player.rocketFuel, 1.5)).times(25).floor()
-      ),
-      rfEff1: computed(() => Decimal.div(player.rocketFuel, 4 / 3).plus(1)),
-      rfEff2: computed(() => Decimal.max(player.rocketFuel, 1).sqrt()),
-      rfEff3: computed(() => Decimal.sub(player.rocketFuel, 1).max(1).cbrt()),
-      rfEff4: computed(() =>
-        Decimal.sub(player.rocketFuel, 2).max(1).log2().plus(1).root(9)
-      ),
-
       effExp: computed(() => {
-        const exp = Decimal.mul(player.rockets, rockets.data.rfEff1.value)
+        const exp = Decimal.mul(player.rockets, rocketFuel.data.eff1.value)
           .plus(1)
           .log10()
-          .times(rockets.data.rfEff4.value);
+          .times(rocketFuel.data.eff4.value);
         return ROCKET_EFF_EXP_SOFTCAP.apply(exp);
       }),
       effMult: computed(() => {
-        let eff = Decimal.mul(player.rockets, rockets.data.rfEff1.value)
+        let eff = Decimal.mul(player.rockets, rocketFuel.data.eff1.value)
           .div(2)
           .plus(1);
         if (eff.gte(10)) eff = eff.times(10).sqrt();
@@ -121,13 +99,5 @@ export const rockets: Feature<RocketData, RocketActions> = addFeature(
 
         signal("reset", 2);
       },
-
-      fuelUp: () => {
-        if (Decimal.lt(player.rockets, rockets.data.rfCost.value)) return;
-
-        player.rockets = Decimal.sub(player.rockets, rockets.data.rfCost.value);
-        player.rocketFuel = Decimal.add(player.rocketFuel, 1);
-      },
     },
-  }
-);
+  });